requirements.txt
numpy>=1.23
torch>=2.0
PyYAML>=6.0

